+++ packages/cspell-types/src/DictionaryDefinition.ts
@@ -48,6 +48,17 @@ export interface DictionaryDefinitionBase {
     noSuggest?: boolean | undefined;
 
     /**
+     * Used to specify the type of dictionary being referenced.
+     * Values:
+     * - `words` - (default) A dictionary containing words.
+     * - `flag-words` - A dictionary containing flag words. Words found in the dictionary will be treated like `flagWords`.
+     * - `ignore-words` - A dictionary containing words to ignore.
+     *    This is the same as setting `noSuggest` to `true`.
+     * @since 10.3.0
+     */
+    kind?: DictionaryKind | undefined;
+
+    /**
      * Some dictionaries may contain forbidden words to prevent compounding from generating
      * words that are not valid in the language. These are often
      * words that are used in other languages or might be generated through compounding.
@@ -148,6 +159,11 @@ export interface DictionaryDefinitionSimple extends DictionaryDefinitionBaseWith
     /**
      * @hide
      */
+    kind?: DictionaryKind | undefined;
+
+    /**
+     * @hide
+     */
     ignoreForbiddenWords?: boolean | undefined;
 
     /**
@@ -193,6 +209,11 @@ interface HiddenFields {
      * @hide
      */
     noSuggest?: undefined;
+
+    /**
+     * @hide
+     */
+    kind?: DictionaryKind | undefined;
 
     /**
      * Not used
@@ -401,3 +422,19 @@ export type DictionaryRef = DictionaryId;
  * @pattern ^(?=!+[^!*,;{}[\]~\n]+$)(?=(.*\w)).+$
  */
 export type DictionaryNegRef = string;
+
+export type DictionaryKindWords = 'words';
+export type DictionaryKindFlagWords = 'flag-words';
+export type DictionaryKindIgnoreWords = 'ignore-words';
+
+/**
+ * The kind of dictionary.
+ *
+ * Used to specify the type of dictionary being referenced.
+ *
+ * Values:
+ * - `words` - A dictionary containing words.
+ * - `flag-words` - A dictionary containing flag words.
+ * - `ignore-words` - A dictionary containing words to ignore.
+ */
+export type DictionaryKind = DictionaryKindWords | DictionaryKindFlagWords | DictionaryKindIgnoreWords;